Takes in the surveys dataset, cleans the data, and then creates a scatterplot comparing two specified columns in the dataset.

The function survey_dotplot first takes in the filepath to the surveys dataset and ensures it is clean using the survey_cleaner function and saving the output to a variable dataframe. Next, a custom ggplot function is created to make a scatterplot comparing the values of x_var and y_var. The titles for each axis are user-specified as the parameters x_title and y_title. Lastly, the function returns the variable containing the saved dotplot survey_plot to be used outside of the bounds of the current function as well as check for proper graphing.

survey_dotplot(filepath, x_var, y_var, x_title, y_title)

Arguments

filepath

The path to the csv file you want to process

x_var

the independent varaible infuencing the y-axis variable

y_var

the dependent variable varying in response to the x-axis variable

x_title

User-specified title for x-axis of plot

y_title

User-specified title for y-axis of plot

Value

survey_plot the scatterplot using the surveys dataset comparing two variables x_var and y_var